requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
black==24.10.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
DateTime==5.5
debugpy==1.8.7
decorator==5.1.1
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.6
fastapi-cli==0.0.6
filelock==3.16.1
fonttools==4.55.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
light-famd==0.0.3
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mypy-extensions==1.0.0
narwhals==1.14.1
nest-asyncio==1.6.0
numpy==2.1.2
outcome==1.3.0.post0
packaging==24.1
panda==0.3.1
pandas==2.2.3
pandas-flavor==0.6.0
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pexpect==4.9.0
pillow==11.0.0
pingouin==0.5.5
platformdirs==4.3.6
prince==0.14.0
prompt_toolkit==3.0.48
protobuf==5.29.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.21.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.26.1
setuptools==75.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==1.4.0
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.4
streamlit==1.41.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tokenize_rt==6.1.0
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
virtualenv==20.27.0
watchdog==6.0.0
watchfiles==1.0.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
xarray==2024.10.0
xgboost==2.1.3
zope.interface==7.2
